% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustTcellExTRECT.R
\name{adjustTcellExTRECT}
\alias{adjustTcellExTRECT}
\title{Adjust TCRA T cell fraction based on tumour purity and copy number}
\usage{
adjustTcellExTRECT(TCRA.out, purity, TCRA.cn, trustPurity = TRUE)
}
\arguments{
\item{TCRA.out}{Output of runTcellExTRECT}

\item{purity}{Tumour purity}

\item{TCRA.cn}{Tumour copynumber at TCRA locus}

\item{trustPurity}{If lower CI of TCRA score > 1 - purity assume tumour copy number is wrong and instead assume 2}
}
\value{
data frame of TCRA scores with purity/copy number adjusted outputs
}
\description{
Adjust TCRA T cell fraction based on tumour purity and copy number
}
