% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineAdj.R
\name{baselineAdj}
\alias{baselineAdj}
\title{Calculate median values of coverage within exons in rolling windows}
\usage{
baselineAdj(vdj_logR_input, vdj_seg, GCcorrect = TRUE)
}
\arguments{
\item{vdj_logR_input}{data frame of positions and coverage values}

\item{vdj_seg}{Segments used for normalisation}

\item{GCcorrect}{Use GC corrected output or not}
}
\value{
Adjusted baseline logR dataframe
}
\description{
Calculate median values of coverage within exons in rolling windows
}
