% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cov_example}
\alias{cov_example}
\title{Example coverage file}
\format{
A data frame of coverage values within TCRA gene
}
\usage{
cov_example
}
\value{
NA
}
\description{
Coverage values of the TCRA gene from the TRACERx100 CRUK0023 R1 sample (hg19 format)
}
\keyword{datasets}
