% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExonDFBed.R
\name{createExonDFBed}
\alias{createExonDFBed}
\title{Function to create a list of exons from bed file of capture regions from exon custom kit}
\usage{
createExonDFBed(pathToBed, hg19or38 = "hg19")
}
\arguments{
\item{pathToBed}{Path to bed file (tsv format assumed)}

\item{hg19_or_38}{hg19 or hg38 version of genome}
}
\value{
data frame of TCRA exons as provided by capture kit
}
\description{
Function to create a list of exons from bed file of capture regions from exon custom kit
}
