% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonsTcellExTRECT.R
\name{exonsTcellExTRECT}
\alias{exonsTcellExTRECT}
\title{Function to run T cell ExTRECT and return data frame}
\usage{
exonsTcellExTRECT(
  vdj.region.df,
  exons.selected,
  vdj.seg,
  hg19_or_38 = "hg19",
  exons.to.use = NULL,
  GC_correct = TRUE,
  median.k = 50,
  median.thresh = 15,
  summariseExons = TRUE
)
}
\arguments{
\item{vdj.region.df}{data frame containing coverage values by position}

\item{exons.selected}{list of exon positions based on exome capture kit used}

\item{vdj.seg}{locations of segments used for calculation of TCRA score}

\item{hg19_or_38}{hg19 or hg38 version of genome}

\item{exons.to.use}{option to manually select which exons to use (defaults to all)}

\item{GC_correct}{whether to use GC correction or not for output}

\item{median.k}{rolling median window}

\item{median.thresh}{threshold to remove exons with low coverage}

\item{summariseExons}{whether or not to summarise by exons (median)}
}
\value{
data frame of TCRA T cell fractions with 95\% CI
}
\description{
Function to run T cell ExTRECT and return data frame
}
