% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exonwindowplot2.R
\name{exonwindowplot2}
\alias{exonwindowplot2}
\title{Method for getting GC content in selected exons}
\usage{
exonwindowplot2(exon.loc, inputseq, extend1)
}
\arguments{
\item{exon.loc}{Location of exons}

\item{inputseq}{Sequence to calculate GC content for}

\item{extend1}{extend exon sequence by specific number of bases}
}
\value{
data frame of GC content for exons
}
\description{
Method for getting GC content in selected exons
}
