% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovFromBam.R
\name{getCovFromBam}
\alias{getCovFromBam}
\title{Function to create TCRA coverage file from bams}
\usage{
getCovFromBam(bamPath, outPath, vdj.seg)
}
\arguments{
\item{bamPath}{Path to bam file}

\item{outPath}{Path to output directory}

\item{vdj.seg}{Location of TCRA file}
}
\description{
Function to create TCRA coverage file from bams
}
