% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLogRdf.R
\name{getLogRdf}
\alias{getLogRdf}
\title{Method for calculating log ratio from single region}
\usage{
getLogRdf(region.df, segs, minCov = 0)
}
\arguments{
\item{region.df}{coverage values with a specific region}

\item{segs}{Location of segments used for normalisation and focal region}

\item{minCov}{Minimum GC coverage required}
}
\value{
Calculated single log ratio based on segments
}
\description{
Method for calculating log ratio from single region
}
