% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVDJFraction_v2.R, R/medianFilter.R
\name{getVDJFraction_upd2}
\alias{getVDJFraction_upd2}
\alias{medianFilter}
\title{Method for calculating final VDJ fraction}
\usage{
getVDJFraction_upd2(
  tumour.logR,
  segs,
  norm.ci.95,
  GC.correct = FALSE,
  ci_type = "simultaneous"
)

medianFilter(x, k)
}
\arguments{
\item{tumour.logR}{log ratio in region}

\item{segs}{Location of segments used for normalisation and focal region}

\item{norm.ci.95}{Adjusted value for 95% CI}

\item{GC.correct}{Whether to use GC corrected value or not}

\item{ci_type}{Method to use for calculation of the confidence interval}
}
\value{
VDJ fraction and upper and lower CI values

VDJ fraction and upper and lower CI values
}
\description{
Method for calculating final VDJ fraction

Method for calculating final VDJ fraction
}
