% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCov.R
\name{loadCov}
\alias{loadCov}
\title{Function to load coverage from getCovFromBam}
\usage{
loadCov(covFile)
}
\arguments{
\item{covFile}{Path to coverage file from getCovFromBam output}
}
\description{
Function to load coverage from getCovFromBam
}
