% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianExonCoverage.R
\name{medianExonCoverage}
\alias{medianExonCoverage}
\title{Calculate median values of coverage within exons in rolling windows}
\usage{
medianExonCoverage(
  vdj.region.df,
  exons.selected,
  median.k = 50,
  median.thresh = 15,
  exons.to.use
)
}
\arguments{
\item{vdj.region.df}{data frame of positions and coverage values}

\item{exons.selected}{Locations of exons}

\item{median.k}{rolling median window}

\item{median.thresh}{threshold to remove exons with low coverage}

\item{exons.to.use}{option to manually select exons}
}
\value{
Data frame of positions and rolling median of coverage values within chosen exons
}
\description{
Calculate median values of coverage within exons in rolling windows
}
