% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianFilter.R
\name{medianFilter}
\alias{medianFilter}
\title{Method for creating running medians}
\usage{
medianFilter(x, k)
}
\arguments{
\item{x}{Vector}

\item{k}{Location of segments used for normalisation and focal region}
}
\value{
vector with running median values
}
\description{
Method for creating running medians
}
