% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTcellExTRECT.R
\name{plotTcellExTRECT}
\alias{plotTcellExTRECT}
\title{Function to plot output from T cell ExTRECT}
\usage{
plotTcellExTRECT(
  vdj.region.df,
  exons.selected,
  vdj.seg,
  hg19_or_38,
  exons.to.use = NULL,
  median.k = 50,
  median.thresh = 15,
  sample_name = "",
  txt.size = 4,
  txt.height = 1.5
)
}
\arguments{
\item{vdj.region.df}{data frame containing coverage values by position}

\item{exons.selected}{list of exon positions based on exome capture kit used}

\item{vdj.seg}{locations of segments used for calculation of TCRA score}

\item{hg19_or_38}{hg19 or hg38 version of genome}

\item{exons.to.use}{option to manually select which exons to use (defaults to all)}

\item{median.k}{rolling median window}

\item{median.thresh}{threshold to remove exons with low coverage}

\item{sample_name}{name of sample run}

\item{txt.size}{size of annotation text}

\item{txt.height}{location on y axis of annotation txt}
}
\value{
data frame of TCRA T cell fractions with 95% CI
}
\description{
Function to plot output from T cell ExTRECT
}
