% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingwindowplot_alt.R
\name{slidingwindowplot_alt}
\alias{slidingwindowplot_alt}
\title{Method for getting GC content in different window sizes}
\usage{
slidingwindowplot_alt(windowsize, inputseq)
}
\arguments{
\item{windowsize}{Length of window to calculate GC content in}

\item{inputseq}{Sequence to calculate GC content for}
}
\value{
data frame of GC content
}
\description{
Method for getting GC content in different window sizes
}
