% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCRA_exons_agilent_v2_hg19}
\alias{TCRA_exons_agilent_v2_hg19}
\title{TCRA exons (hg19)}
\format{
A data frame of exon locations
}
\usage{
TCRA_exons_agilent_v2_hg19
}
\value{
NA
}
\description{
Exons for Agilent v2 capture kit within TCRA loci, based
on Agilent v4/5 but with exons showing bias removed (hg19)
}
\keyword{datasets}
