% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCRA_exons_hg38}
\alias{TCRA_exons_hg38}
\title{TCRA exons (hg38)}
\format{
A data frame of exon locations
}
\usage{
TCRA_exons_hg38
}
\value{
NA
}
\description{
Exons used in Agilent v4/5 capture kit within TCRA loci, extracted
from bed file (hg38 version)
}
\keyword{datasets}
