% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCRA_exons_nimblegen_hg19}
\alias{TCRA_exons_nimblegen_hg19}
\title{TCRA exons nimblegen (hg19)}
\format{
A data frame of exon locations
}
\usage{
TCRA_exons_nimblegen_hg19
}
\value{
NA
}
\description{
Exons used in Nimblegen capture kit within TCRA loci, extracted
from bed file (hg19 version)
}
\keyword{datasets}
