% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCRA_exons_nimblegen_hg38}
\alias{TCRA_exons_nimblegen_hg38}
\title{TCRA exons from Nimblegen SeqCap_EZ_Exome_v3_hg19_capture_targets.bed (hg19)}
\format{
A data frame of exon locations
}
\usage{
TCRA_exons_nimblegen_hg38
}
\value{
NA
}
\description{
Exons used in Nimblegen SeqCap EZ Exome v3 capture kit within TCRA loci, extracted
from bed file (hg19 version)
}
\keyword{datasets}
