% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCRA_fasta}
\alias{TCRA_fasta}
\title{TCRA fasta file}
\format{
A list containing a SeqFastadna object
}
\usage{
TCRA_fasta
}
\value{
NA
}
\description{
Sequence around the TCRA locus from fasta file
chr14:22000000-23110000 (hg19)
chr14:???(hg38)
}
\keyword{datasets}
