% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcra_seg_hg19}
\alias{tcra_seg_hg19}
\title{TCRA segments file (hg19)}
\format{
A list containing regions used to run T cell ExTRECT
}
\usage{
tcra_seg_hg19
}
\value{
NA
}
\description{
Segments used for running T cell ExTRECT, defining TCRA loci,
local regions used for normalisation and region of maximum VDJ
recombination (hg19 version)
}
\keyword{datasets}
