% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TCRA_segments}
\alias{TCRA_segments}
\title{TCRA segments}
\format{
A data frame of TCRA segments
}
\usage{
TCRA_segments
}
\value{
NA
}
\description{
Data frame of the VDJ gene segments within the TCRA locu
}
\keyword{datasets}
