% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCcorrect.R
\name{GCcorrect}
\alias{GCcorrect}
\title{Method for correcting for GC biases}
\usage{
GCcorrect(
  tumour.logR,
  exons,
  exonList,
  gene.fasta,
  hg19_or38 = "hg19",
  sliding = 1000
)
}
\arguments{
\item{tumour.logR}{LogR values}

\item{exons}{Location of exons}

\item{exonList}{GC content of exons}

\item{gene.fasta}{offset for FASTA file, e.g 21999999 for TCRA in hg19}
}
\value{
data frame of GC correct logR values
}
\description{
Method for correcting for GC biases
}
