% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVDJFraction.R
\name{getVDJFraction}
\alias{getVDJFraction}
\title{Original method for calculating final VDJ fraction}
\usage{
getVDJFraction(tumour.logR, segs, GC.correct = FALSE)
}
\arguments{
\item{tumour.logR}{log ratio in region}

\item{segs}{Location of segments used for normalisation and focal region}

\item{GC.correct}{Whether to use GC corrected value or not}
}
\value{
VDJ fraction and upper and lower CI values
}
\description{
Original method for calculating final VDJ fraction
}
